Remove aio-pika and psutil
aio-pika was used in the RabbitMQ cog (added in https://github.com/python-discord/bot/pull/90) which has since been deleted. It is no longer used.

psutil was used to automatically determine how many threads to use when testing. However, this package is no longer maintained, so I have removed it and hardcoded the number of threads to 8.
An error is not raised it the number of threads used exceeds the number of actual threads available.